Identify R functions best matching a given input string

Function matching is only available for Only applies to functions from the corpus of rOpenSci packages.

Usage

pkgmatch_similar_fns(input, embeddings = NULL, n = 5L, browse = FALSE)

Arguments

input

A text string.

embeddings

Large Language Model embeddings for all rOpenSci packages, generated from pkgmatch_embeddings_from_pkgs. If not provided, pre-generated embeddings will be downloaded and stored in a local cache directory.

n

When the result of this function is printed to screen, the top n packages will be displayed.

browse

If TRUE, automatically open webpages of the top n matches in local browser.

Value

A character vector of function names in the form "::".

See also

Examples

if (FALSE) { # \dontrun{
input <- "Process raster satellite images"
p <- pkgmatch_similar_fns (input)
p # Default print method, lists 5 best matching packages
head (p) # Shows first 5 rows of full `data.frame` object
} # }